# docs/openapi.yaml
openapi: 3.0.3
info:
  title: STEPQUEST Backend API
  version: 0.2.0
  description: API contract for the current STEPQUEST MVP backend and legacy account flow.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    CreateVisionGoalRequest:
      type: object
      required: [title, targetDate]
      properties:
        title:
          type: string
        description:
          type: string
        targetDate:
          type: string
          format: date
    GenerateWeeklyPlanRequest:
      type: object
      properties:
        weeks:
          type: integer
        targetCount:
          type: integer
    GenerateMicroActionsRequest:
      type: object
      properties:
        count:
          type: integer
    VisionGoal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        targetDate:
          type: string
          format: date
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          nullable: true
    WeeklyMission:
      type: object
      properties:
        id:
          type: integer
        visionGoalId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        weekStartDate:
          type: string
        weekEndDate:
          type: string
        targetCount:
          type: integer
        completedCount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          nullable: true
    MicroAction:
      type: object
      properties:
        id:
          type: integer
        weeklyMissionId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        estimatedSeconds:
          type: integer
        difficulty:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          nullable: true
    MicroCompleteResponse:
      type: object
      properties:
        microActionId:
          type: integer
        grantedGoalCoin:
          type: integer
        dailyGoalCoinEarned:
          type: integer
        currentStreakDays:
          type: integer
        consistencyScore:
          type: integer
    MicroShrinkResponse:
      type: object
      properties:
        microActionId:
          type: integer
        replacementSteps:
          type: array
          items:
            $ref: '#/components/schemas/MicroAction'
        firstStep:
          $ref: '#/components/schemas/MicroAction'
        reason:
          type: string
    StepQuestCreateGoalRequest:
      type: object
      required: [title, burdenLevel]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [auto, study, work, writing, cleaning, exercise, wake, sleep, life_admin, relationship]
        burdenLevel:
          type: integer
          minimum: 1
          maximum: 4
        energyLevel:
          type: string
          enum: [low, medium, high]
        targetAt:
          type: string
          format: date-time
        recurrenceRule:
          type: string
          enum: [daily, weekly]
        location:
          type: string
        availableMinutes:
          type: integer
          minimum: 1
          maximum: 180
        obstacle:
          type: string
          enum: [too_big, no_material, unclear, tired, wrong_place, not_now, forgot, anxious]
...
    StepQuestAnyResponse:
      type: object
      additionalProperties: true
...
paths:
  /stepquest/goals:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Create a STEPQUEST goal, quest chain, and first active micro step.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepQuestCreateGoalRequest'
      responses:
        '201':
          description: Created STEPQUEST goal state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
...
  /goals/vision:
    get:
      tags: [Goals]
      security:
        - bearerAuth: []
      summary: List vision goals.
      responses:
        '200':
          description: Vision goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisionGoal'
    post:
      tags: [Goals]
      security:
        - bearerAuth: []
      summary: Create a vision goal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisionGoalRequest'
      responses:
        '200':
          description: Vision created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionGoal'
  /goals/vision/{id}/weekly-plan:
    post:
      tags: [Goals]
      security:
        - bearerAuth: []
      summary: Generate weekly missions.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateWeeklyPlanRequest'
      responses:
        '200':
          description: Weekly missions created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeeklyMission'
  /goals/weekly:
    get:
      tags: [Goals]
      security:
        - bearerAuth: []
      summary: List weekly missions.
      parameters:
        - name: visionId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Weekly missions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeeklyMission'
  /goals/weekly/{id}/micro:
    get:
      tags: [Goals]
      security:
        - bearerAuth: []
      summary: List micro actions for a weekly mission.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Micro actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MicroAction'
  /goals/weekly/{id}/micro-generate:
    post:
      tags: [Goals]
      security:
        - bearerAuth: []
      summary: Generate micro actions.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateMicroActionsRequest'
      responses:
        '200':
          description: Micro actions created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MicroAction'
  /actions/next-micro:
    get:
      tags: [Actions]
      security:
        - bearerAuth: []
      summary: Get the smallest open micro action.
      responses:
        '200':
          description: Next micro action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroAction'
  /actions/micro/{id}/complete:
    post:
      tags: [Actions]
      security:
        - bearerAuth: []
      summary: Complete a micro action.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Completion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroCompleteResponse'
  /actions/micro/{id}/shrink:
    post:
      tags: [Actions]
      security:
        - bearerAuth: []
      summary: Replace a too-large micro action with smaller steps.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Shrink result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroShrinkResponse'
...